make change to new bonus
rename main.c to 2022-sharing-session and add new single_crud.c
